join all the threads
